# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CustomerVoice API (Sprint 1 Scaffold)
  version: 0.1.0
servers:
  - url: http://localhost:4000/api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
  /workspaces/{workspaceId}/members:
    get:
      summary: List workspace members
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Member list
  /workspaces/{workspaceId}/members/invite:
    post:
      summary: Invite member
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Member invited
  /workspaces/{workspaceId}/members/{userId}/role:
    patch:
      summary: Update member role
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Role updated
  /workspaces/{workspaceId}/members/{userId}:
    delete:
      summary: Deactivate member
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Member deactivated
  /workspaces/{workspaceId}/audit-events:
    get:
      summary: List audit events
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Audit event list